split tests into 2 files
